v1_archive/requirements.txt
openai>=1.0.0
anthropic>=0.18.0
google-genai>=0.1.0
httpx>=0.25.0
pydantic>=2.0.0
python-dotenv>=1.0.0
rich>=13.0.0
asyncio-throttle>=1.0.0
